Update fpgui.fr.po
I cannot agree with last change. Changes for short text in case of month names add a point termination and even additional letters which are in contradiction with the need of shortened month names. It is strange that the same was not done for week day names. There is also some incorrect sentences like the one in line 122 which could be "Cliquer sur la pipette et maintenir le bouton => relâcher pour copier la couleur", eventually.